(MAINT) Add initial implementation
This commit implements the hugo-module-update action and provides
documentation for it. In this initial implementation, the composite
action will install hugo, call `hugo mod get -u ./..`. against the
project, and capture the output.

If the `recursive` parameter is set to `false`, it will run
`hugo mod get -u` instead.

If the `no-install` parameter is set to `true`, it will skip
installing hugo entirely.

The output is transformed into a JSON string so that the multiline
updates can be properly interpolated in a subsequent step.